+++ src/index.d.ts
@@ -371,6 +371,8 @@ import {
 } from './services/user/management.js';
 
 import {
+	fetchHasActivePlatformSubscription,
+	fetchLastSubscriptionPlatform,
 	fetchMemberships,
 	fetchRechargeTokens,
 	getUpgradePrice,
@@ -379,6 +381,7 @@ import {
 } from './services/user/memberships.ts';
 
 import {
+	deleteAllNotifications,
 	deleteNotification,
 	fetchLiveEventPollingState,
 	fetchNotificationSettings,
@@ -480,6 +483,7 @@ declare module 'musora-content-services' {
 		createTestUser,
 		createThread,
 		deleteAccount,
+		deleteAllNotifications,
 		deleteComment,
 		deleteForumCategory,
 		deleteItemsFromPlaylist,
@@ -525,10 +529,16 @@ declare module 'musora-content-services' {
 		fetchGenreBySlug,
 		fetchGenreLessons,
 		fetchGenres,
+<<<<<<< fix/TP-1105_playlist-progress-for-homepage-row
+=======
+		fetchHasActivePlatformSubscription,
+		fetchHierarchy,
+>>>>>>> main
 		fetchInstructorBySlug,
 		fetchInstructorLessons,
 		fetchInstructors,
 		fetchInterests,
+		fetchLastSubscriptionPlatform,
 		fetchLatestThreads,
 		fetchLearningPathLessons,
 		fetchLearningPathProgressCheckLessons,